Template ref "ProgressPosition" Exposed
Template ref "ProgressPosition" has been exposed to parent through the defineExpose() method. While creating a video player with it, volume controller was successfully loaded and working. But we couldn't set the video/audio progress whilethe file plays. Even if we were setting the "--progress-position" property via setProperty() method it was not working properly. That's why we simply exposed "ProgressPosition" and now we can easily get access to it from the parent and update as the file continues playing. Thanks to Rejown Ahmed who made this possible.